% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{$.market_tbl}
\alias{$.market_tbl}
\title{Extract column from market_tbl using $}
\usage{
\method{$}{market_tbl}(x, name)
}
\arguments{
\item{x}{A market_tbl object}

\item{name}{Column name}
}
\description{
Extract column from market_tbl using $
}
