% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{dplyr_reconstruct.market_tbl}
\alias{dplyr_reconstruct.market_tbl}
\title{Reconstruct market_tbl after dplyr operations}
\usage{
dplyr_reconstruct.market_tbl(data, template)
}
\arguments{
\item{data}{Result of dplyr operation}

\item{template}{Original market_tbl template}
}
\description{
Reconstruct market_tbl after dplyr operations
}
