% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_market_tbl}
\alias{is_market_tbl}
\title{Check if Market Data is Valid}
\usage{
is_market_tbl(data)
}
\arguments{
\item{data}{Object to check}
}
\value{
Logical: TRUE if valid market data
}
\description{
Quick check if data is a valid market_tbl object.
}
