% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{[.market_tbl}
\alias{[.market_tbl}
\title{Subset market_tbl}
\usage{
\method{[}{market_tbl}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A market_tbl object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{drop}{Logical, drop dimensions}
}
\description{
Subset market_tbl
}
