% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{[[.market_tbl}
\alias{[[.market_tbl}
\title{Extract column from market_tbl}
\usage{
\method{[[}{market_tbl}(x, i, exact = TRUE)
}
\arguments{
\item{x}{A market_tbl object}

\item{i}{Column index or name}

\item{exact}{Logical, exact matching}
}
\description{
Extract column from market_tbl
}
